initial: Create a bot that can launch clusters via ProwJobs
OpenShift CI runs and launches clusters to test e2e tests, and also
offers per-PR cluster setup for quick debugging. Extend this to general
use by accepting slack requests to spin up and manage an OpenShift
cluster. This reuses the same infrastructure as the clusters we run
to verify a release payload.

Allow only a max number of clusters at once. Launch a prow job
when a user requests via the `launch` command. Track that cluster
until it expires. The clusters last for 2 hours and then terminate
and are cleaned up. Artifacts go to the appropriate log locations
on GCS.