Bot should reply with link to prow URL for non-launch jobs
Non-launch jobs are only accessible via their prow URL and so are
difficult to find if we don't respond with the link. Capture the
user's original command and report that back in both list and response
so that it's easier to associate input and output.

Fix debug statements to allow more consistency when debugging the
job across multiple tests.