During setup, attempt to set the `BRANCH` variable
`BRANCH` is used to locate images that are not part of the payload
that may be necessary for UPI installs (or other testing). Try to
detect the version from the base ref or the current payload version
(either PR or release job) in order to set it. This may have some
gaps for jobs that run custom images which will have to be fixed
later.