Allow test jobs to be launched from `test <type>`
The bot looks them up in the config spec of the provided job - if
they match by name then the test will proceed (the jobs must be by
platform since the mount parameters differ).

The list of test types is inlined but checked against the job for
future extensibility or testing new types.