Switching image resolution over to the app.ci cluster
We recently relocated the release-controller to the app.ci cluster.
This meant that the imagestreams that the ci-chat-bot relies on,
no longer exist at the location it's currently monitoring.  This PR
adds a new option to override the imagestream client, if necessary,
to point at any cluster to read the release imagestreams.