Accept a release image when upgrading correctly
We unset the release images before calling the final step of the
ci-operator job, which means the upgrade doesn't occur from whichever
argument is explicitly an image. Set the appropriate variables for
non-PR content (anything we build we need to build a payload for)
so that we use that release image.