Use new registry_override field when promoting build jobs
Now that promotion uses mirroring exclusively, we need to prepare
a pull secret using the pod's service account for use when promoting
from the build namespace (<ns>-N) to the main namespace (<ns>). We
need to know the public URL because the promotion step uses the public
facing URL of the cluster for the input image.

Long term chat bot also needs to change to load the appropriate cluster
config from the build-kubeconfig based on the job, then spawn clients
(the current behavior only accidentally works as long as the job and
chat bot config points to the same dir).