Clone repos in build jobs so that ci-operator can read repo config
ci-operator added `.ci-operator.yaml` file support, which requires
jobs that build PRs to clone the repo so that config is available.
Set skip_cloning: false, then set a REPO_PATH for each sub build.
The "from" and "to" parts (inputs 0 and 1) have different root
dirs so that we can have two different PRs for the same repo, which
required us to merge a fix to clonerefs.