Commandline support for ProwJob generation
The ART team has requested a way to automate their process
of manually firing off upgrade jobs, via the chat-bot, when
new releases are accepted.  We came up with a solution that
involves using the chat-bot, as a command-line tool, to
generate the necessary ProwJob's and then ART will submit
the jobs accordingly.

The majority of this PR is just refactoring the logic to
allow the generated ProwJob payload's to be exported instead
of submitted to the build cluster. I have added support for
all the existing commands, that generate ProwJobs, using the
same logic for either path (slack or commandline).

A sample commandline execution looks like this:
```
./ci-chat-bot --build-cluster-kubeconfigs-location /var/build-cluster-kubeconfigs --prow-config=/etc/config/config.yaml --job-config=/etc/job-config --command 'test upgrade 4.7.0 4.7.14' --output upgrade.json
```