manager: support multiple config-types
This PR adds support for loading the "modern" config-type and falling
back to "legacy" if a matching config is not found in the modern job.
This chat bot will automatically identify which parameters need to be
passed via environement variables and which ones require a separate
launch config in a platform dependent manner.

This PR also imports the ci-tools repo to allow easier inspection and
modification of the ci-operator configs.