prow,manager: add support for upgrade jobs
This PR adds the ability to run upgrade jobs on the modern config type.
At the time of writing this commit message, the `e2e-upgrade-all` test
type is unable to be accurately run using the step registry as the step
registry's script hardcodes the parallel conformance suite for the
`upgrade-conformance` test type, while the exisiting `e2e-upgrade-all`
chatbot test uses the full conformance suite. In this PR, the chatbot
will fall back to templates for that test type until we can reconcile
this difference.