Add new workflow-launch command
This PR adds a new `workflow-launch` command to the `ci-chat-bot`. This
command allows users to run any workflow they want via the `ci-chat-bot`
as long as that workflow is defined in a simple config file in
`openshift/release`. The workflows config file maps workflows to their
correct platform (and potentially architecture) as well as any
`base_images` that the workflow may require.