slack: Make it easier to self-help workflow-launch failures
By listing known workloads and pointing folks at where they can
(hopefully) add more.  Assuming the config comes from the release repo
is brittle, but since we're probably the only folks running this
thing, it may help.  I guess we could also add an explicit config knob
for "where to suggest changes to this config", but I'm punting on that
for now.