manager: use first bundle from multi-bundle config
This PR allows users to launch an operator in repos that define more
than one bundle in their configuration. It will take the first bundle in
the list and inform the user what bundle name/config is being used.
This should only be a temporary measure; in the future, users should be
able to specify which bundle they want from the ones defined in the
config.