Makefile: use openshift machinery
This PR adds support for using the openshift build machinery for the
Makefile. It also moves the `main.go` and `slack.go` files out of the
root to a `cmd` subdirectory and adds versioning that is output to
stdout when the chat-bot is run. The following new make targets can be
added to the ci-operator config tests once this PR is in:

- make verify
- make verify-golint (use golangci-lint image that ci-tools uses)
- make validate-vendor

Once the ci-operator config is updated, we will also be able to remove
the old hack/linter.sh, as the verify and validate-vendor targets
replace it.