slack: make hypershift-hosted default for supported versions
This PR makes the `hypershift-hosted` platform the default for all
versions that `hypershift` on the hive cluster supports. It also adds a
global variable to define the default openshift release version used by
`ci-chat-bot`, makes the `HypershiftSupportedVersions` variable global
to make it simpler to access across packages, adds a launch message for
cluster launched using `hypershift-hosted`, and removes the estimated
launch time from the cluster creation message as that is generally
inaccurate now.