manager: add hypershift supported-versions handler
This PR adds a new goroutine to the job manager that maintains a list of
supported versions for our instance of hypershift on the DPTP hive
cluster based on the `supported-versions` configmap in the `hypershift`
namespace.

The `hypershiftSupportedVersions` variable is currently being created in
the `main.go` as a future PR will also provide it to the slack bot for
intelligent defaulting to hypershift based on the provided release
version.