*: add support for rosa
This PR adds a `rosa create <version>` command that allows a user whose
ID is in a whitelist to launch clusters in the ROSA environment. The
`auth` and `done` commands also work on ROSA clusters and if a user has
a ROSA cluster, the cluster will show up in the `list` command. The
chat-bot keeps passwords that it creates in a secret. When a cluster ID
in the secret no longer appears in the list of clusters returned by
ROSA, commands to clean up created account-roles and oidc-provider are
run.

Currently, the timeout is the same as normal job launches (3 hours + 45
minutes for launching).