rosa: separate console wait from auth
This PR makes the console wait a separate function than the `auth` wait
and changes behavior to initially notify a user once auth is available,
instead of trying to wait for console as well. It also extends the wait
time for console to 20 minutes after auth, as it was hitting the 10
minute timeout frequently.