manager: stop spamming failed rosa clusters
This PR keeps an in-memory log of cluster IDs that the chat-bot has
reported failures for. In some cases, a cluster will go back and forth
between states (specifically "Uninstalling" and "Error"), which causes
the error message to be spammed to the user. Using a set to track this
prevents the error from being sent to a user more than once per instance
of cluster-bot.