Structured help page formatting
The current help page is a little hard to parse as there's no hierarchy
of information. Attempt to improve this by making use of headers, in the
absence of other structure elements (e.g. horizontal lines) in Slack's
Markdown, to produce a manpage-like format.

Before:

  `rosa` `lookup` `version`

  _Find openshift version(s) with provided prefix that is supported in ROSA._

  > *Example:* _rosa lookup 4.15_

After:

  | *rosa lookup*
  *Usage*
  ```
  rosa lookup version
  ```
  *Description*
  Find openshift version(s) with provided prefix that is supported in ROSA.
  *Example*
  ```
  rosa lookup 4.15
  ```

(you'll likely want to look at those with markdown rendering to make
sense of them)

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>