manager: support nightlies and image builds for MCE
This PR adds support for using prowjobs to create images pointing to nightlies
or built from PRs that are able to be pulled without authentication using the
`ipi-install-rbac` multistage test step. When a nightly or image build is
requested, a prowjob is started to build the image and run the required step,
and once it is complete, an imageset pointing to the release is created and a
`ClusterDeployment` pointing to the imageset is created for the corresponding
`ManagedCluster`. If an error occurs during the prow step, the user is sent an
error message and the existing MCE resources are deleted.