Bump the Max ROSA Age up to 8 hours
Using a ClusterBot ROSA cluster for a day can get frustrating
when the cluster is torn down towards the end of the day.
Bump up the Max age so that it is more likely to span the length
of an entire working day. Leave the default as is so that this
change is less likely to suddenly impact the overall resource
usage footprint of clusterbot.

Signed-off-by: Derek Higgins <derekh@redhat.com>