*: add interactive MCE cluster management (#574)
* *: add interactive MCE cluster management

This PR adds interactive elements for managing MCE cluster. It also updates and
completes existing cluster management modals and removes duplicate code blocks
to improve maintainability. These changes will require now require the
`im.write` permission in Slack in order to allow the `auth` functionality to
function correctly.

* apphome: redesign app home

Improves App Home and makes it more dynamic and far easier to maintain in the
future.

This also adds an option to disable rosa for local running since the timeout
for both rosa and aws logins is incredibly short and unnecessary for most
testing.

* modals: use reverse semantic sorting for major.minor list

* modals: fix a variety of image selection issues

This fixes a bunch of issues discovered for the image selection:
- Make selection for ci and nightly work
- Require a selection between stream/latest/custom
- Make Major.Minor and Version required
- Don't display ci stream if it doesn't exist (like in multiarch)
- Ignore cancelled prowjobs when finding user's CI cluster
- Fix building PRs with specified image version

* modals: move launch consts and version detection to common location

Move all `Launch...` consts shared between Launch and MCE_Create to
a common `types.go` in the parent `modals` package.

Create a `GetVersion` function to identify the requested version from
the callback data that can be used by both Launch and MCE_Create.