modals: split cluster lists into multiple blocks
This works around the 3000 character limit for a markdown element by using
multiple sections. There is a maximum block limit of 100 blocks and the maximum
number of blocks for the main `list` modal is 81. This also removes the `rosa`
and `jobs` listing from the modal list, as that could result in >100 block
elements. The `rosa` clusters and non-cluster `jobs` will still be listed in
the message based `list` command.